+++ internal/storage/repos/repo_currencies/currencies.sql.go
@@ -13,7 +13,7 @@ import (
 )
 
 const getCurrenciesByBlockchain = `-- name: GetCurrenciesByBlockchain :many
-SELECT id, code, name, precision, is_fiat, blockchain, contract_address, withdrawal_min_balance, has_balance, status, sort_order, min_confirmation, created_at, updated_at, is_stablecoin, currency_label, token_label, is_native
+SELECT id, code, name, precision, is_fiat, blockchain, contract_address, withdrawal_min_balance, has_balance, status, sort_order, min_confirmation, created_at, updated_at, is_stablecoin, currency_label, token_label, is_native, is_new_store_default, order_idx
 FROM currencies
 WHERE blockchain = $1
 `
@@ -46,6 +46,8 @@ func (q *Queries) GetCurrenciesByBlockchain(ctx context.Context, blockchain *mod
 			&i.CurrencyLabel,
 			&i.TokenLabel,
 			&i.IsNative,
+			&i.IsNewStoreDefault,
+			&i.OrderIdx,
 		); err != nil {
 			return nil, err
 		}
@@ -58,7 +60,7 @@ func (q *Queries) GetCurrenciesByBlockchain(ctx context.Context, blockchain *mod
 }
 
 const getCurrenciesEnabled = `-- name: GetCurrenciesEnabled :many
-SELECT id, code, name, precision, is_fiat, blockchain, contract_address, withdrawal_min_balance, has_balance, status, sort_order, min_confirmation, created_at, updated_at, is_stablecoin, currency_label, token_label, is_native
+SELECT id, code, name, precision, is_fiat, blockchain, contract_address, withdrawal_min_balance, has_balance, status, sort_order, min_confirmation, created_at, updated_at, is_stablecoin, currency_label, token_label, is_native, is_new_store_default, order_idx
 FROM currencies
 WHERE status = true
 ORDER BY blockchain, code
@@ -92,6 +94,8 @@ func (q *Queries) GetCurrenciesEnabled(ctx context.Context) ([]*models.Currency,
 			&i.CurrencyLabel,
 			&i.TokenLabel,
 			&i.IsNative,
+			&i.IsNewStoreDefault,
+			&i.OrderIdx,
 		); err != nil {
 			return nil, err
 		}
@@ -104,7 +108,7 @@ func (q *Queries) GetCurrenciesEnabled(ctx context.Context) ([]*models.Currency,
 }
 
 const getCurrenciesHasBalance = `-- name: GetCurrenciesHasBalance :many
-SELECT id, code, name, precision, is_fiat, blockchain, contract_address, withdrawal_min_balance, has_balance, status, sort_order, min_confirmation, created_at, updated_at, is_stablecoin, currency_label, token_label, is_native
+SELECT id, code, name, precision, is_fiat, blockchain, contract_address, withdrawal_min_balance, has_balance, status, sort_order, min_confirmation, created_at, updated_at, is_stablecoin, currency_label, token_label, is_native, is_new_store_default, order_idx
 FROM currencies
 where status = true
   and has_balance = true
@@ -138,6 +142,8 @@ func (q *Queries) GetCurrenciesHasBalance(ctx context.Context) ([]*models.Curren
 			&i.CurrencyLabel,
 			&i.TokenLabel,
 			&i.IsNative,
+			&i.IsNewStoreDefault,
+			&i.OrderIdx,
 		); err != nil {
 			return nil, err
 		}
@@ -150,7 +156,7 @@ func (q *Queries) GetCurrenciesHasBalance(ctx context.Context) ([]*models.Curren
 }
 
 const getCurrencyByBlockchainAndContract = `-- name: GetCurrencyByBlockchainAndContract :one
-SELECT id, code, name, precision, is_fiat, blockchain, contract_address, withdrawal_min_balance, has_balance, status, sort_order, min_confirmation, created_at, updated_at, is_stablecoin, currency_label, token_label, is_native
+SELECT id, code, name, precision, is_fiat, blockchain, contract_address, withdrawal_min_balance, has_balance, status, sort_order, min_confirmation, created_at, updated_at, is_stablecoin, currency_label, token_label, is_native, is_new_store_default, order_idx
 FROM currencies
 WHERE blockchain = $2
   AND (contract_address = $1 OR (contract_address IS NULL AND code = UPPER($1)))
@@ -185,12 +191,14 @@ func (q *Queries) GetCurrencyByBlockchainAndContract(ctx context.Context, arg Ge
 		&i.CurrencyLabel,
 		&i.TokenLabel,
 		&i.IsNative,
+		&i.IsNewStoreDefault,
+		&i.OrderIdx,
 	)
 	return &i, err
 }
 
 const getCurrencyWithBalanceByBlockchain = `-- name: GetCurrencyWithBalanceByBlockchain :one
-SELECT id, code, name, precision, is_fiat, blockchain, contract_address, withdrawal_min_balance, has_balance, status, sort_order, min_confirmation, created_at, updated_at, is_stablecoin, currency_label, token_label, is_native
+SELECT id, code, name, precision, is_fiat, blockchain, contract_address, withdrawal_min_balance, has_balance, status, sort_order, min_confirmation, created_at, updated_at, is_stablecoin, currency_label, token_label, is_native, is_new_store_default, order_idx
 FROM currencies
 WHERE blockchain = $1
   AND has_balance = true
@@ -219,12 +227,14 @@ func (q *Queries) GetCurrencyWithBalanceByBlockchain(ctx context.Context, blockc
 		&i.CurrencyLabel,
 		&i.TokenLabel,
 		&i.IsNative,
+		&i.IsNewStoreDefault,
+		&i.OrderIdx,
 	)
 	return &i, err
 }
 
 const getEnabledCurrencyByCode = `-- name: GetEnabledCurrencyByCode :one
-SELECT id, code, name, precision, is_fiat, blockchain, contract_address, withdrawal_min_balance, has_balance, status, sort_order, min_confirmation, created_at, updated_at, is_stablecoin, currency_label, token_label, is_native
+SELECT id, code, name, precision, is_fiat, blockchain, contract_address, withdrawal_min_balance, has_balance, status, sort_order, min_confirmation, created_at, updated_at, is_stablecoin, currency_label, token_label, is_native, is_new_store_default, order_idx
 FROM currencies
 WHERE code = $1
   AND status = true
@@ -259,12 +269,14 @@ func (q *Queries) GetEnabledCurrencyByCode(ctx context.Context, arg GetEnabledCu
 		&i.CurrencyLabel,
 		&i.TokenLabel,
 		&i.IsNative,
+		&i.IsNewStoreDefault,
+		&i.OrderIdx,
 	)
 	return &i, err
 }
 
 const getEnabledCurrencyById = `-- name: GetEnabledCurrencyById :one
-SELECT id, code, name, precision, is_fiat, blockchain, contract_address, withdrawal_min_balance, has_balance, status, sort_order, min_confirmation, created_at, updated_at, is_stablecoin, currency_label, token_label, is_native
+SELECT id, code, name, precision, is_fiat, blockchain, contract_address, withdrawal_min_balance, has_balance, status, sort_order, min_confirmation, created_at, updated_at, is_stablecoin, currency_label, token_label, is_native, is_new_store_default, order_idx
 FROM currencies
 where id = $1 and status = true
 limit 1
@@ -292,6 +304,8 @@ func (q *Queries) GetEnabledCurrencyById(ctx context.Context, id string) (*model
 		&i.CurrencyLabel,
 		&i.TokenLabel,
 		&i.IsNative,
+		&i.IsNewStoreDefault,
+		&i.OrderIdx,
 	)
 	return &i, err
 }