+++ internal/storage/repos/repo_currencies/currencies_gen.sql.go
@@ -14,9 +14,9 @@ import (
 )
 
 const create = `-- name: Create :one
-INSERT INTO currencies (id, code, name, precision, is_fiat, blockchain, contract_address, withdrawal_min_balance, has_balance, status, sort_order, min_confirmation, created_at, is_stablecoin, currency_label, token_label, is_native)
-	VALUES ($1, $2, $3, $4, $5, $6, $7, $8, $9, $10, $11, $12, $13, $14, $15, $16, $17)
-	RETURNING id, code, name, precision, is_fiat, blockchain, contract_address, withdrawal_min_balance, has_balance, status, sort_order, min_confirmation, created_at, updated_at, is_stablecoin, currency_label, token_label, is_native
+INSERT INTO currencies (id, code, name, precision, is_fiat, blockchain, contract_address, withdrawal_min_balance, has_balance, status, sort_order, min_confirmation, created_at, is_stablecoin, currency_label, token_label, is_native, is_new_store_default, order_idx)
+	VALUES ($1, $2, $3, $4, $5, $6, $7, $8, $9, $10, $11, $12, $13, $14, $15, $16, $17, $18, $19)
+	RETURNING id, code, name, precision, is_fiat, blockchain, contract_address, withdrawal_min_balance, has_balance, status, sort_order, min_confirmation, created_at, updated_at, is_stablecoin, currency_label, token_label, is_native, is_new_store_default, order_idx
 `
 
 type CreateParams struct {
@@ -37,6 +37,8 @@ type CreateParams struct {
 	CurrencyLabel        pgtype.Text        `db:"currency_label" json:"currency_label"`
 	TokenLabel           pgtype.Text        `db:"token_label" json:"token_label"`
 	IsNative             bool               `db:"is_native" json:"is_native"`
+	IsNewStoreDefault    bool               `db:"is_new_store_default" json:"is_new_store_default"`
+	OrderIdx             int64              `db:"order_idx" json:"order_idx"`
 }
 
 func (q *Queries) Create(ctx context.Context, arg CreateParams) (*models.Currency, error) {
@@ -58,6 +60,8 @@ func (q *Queries) Create(ctx context.Context, arg CreateParams) (*models.Currenc
 		arg.CurrencyLabel,
 		arg.TokenLabel,
 		arg.IsNative,
+		arg.IsNewStoreDefault,
+		arg.OrderIdx,
 	)
 	var i models.Currency
 	err := row.Scan(
@@ -79,12 +83,14 @@ func (q *Queries) Create(ctx context.Context, arg CreateParams) (*models.Currenc
 		&i.CurrencyLabel,
 		&i.TokenLabel,
 		&i.IsNative,
+		&i.IsNewStoreDefault,
+		&i.OrderIdx,
 	)
 	return &i, err
 }
 
 const getAll = `-- name: GetAll :many
-SELECT id, code, name, precision, is_fiat, blockchain, contract_address, withdrawal_min_balance, has_balance, status, sort_order, min_confirmation, created_at, updated_at, is_stablecoin, currency_label, token_label, is_native FROM currencies
+SELECT id, code, name, precision, is_fiat, blockchain, contract_address, withdrawal_min_balance, has_balance, status, sort_order, min_confirmation, created_at, updated_at, is_stablecoin, currency_label, token_label, is_native, is_new_store_default, order_idx FROM currencies
 `
 
 func (q *Queries) GetAll(ctx context.Context) ([]*models.Currency, error) {
@@ -115,6 +121,8 @@ func (q *Queries) GetAll(ctx context.Context) ([]*models.Currency, error) {
 			&i.CurrencyLabel,
 			&i.TokenLabel,
 			&i.IsNative,
+			&i.IsNewStoreDefault,
+			&i.OrderIdx,
 		); err != nil {
 			return nil, err
 		}
@@ -127,7 +135,7 @@ func (q *Queries) GetAll(ctx context.Context) ([]*models.Currency, error) {
 }
 
 const getByID = `-- name: GetByID :one
-SELECT id, code, name, precision, is_fiat, blockchain, contract_address, withdrawal_min_balance, has_balance, status, sort_order, min_confirmation, created_at, updated_at, is_stablecoin, currency_label, token_label, is_native FROM currencies WHERE id=$1 LIMIT 1
+SELECT id, code, name, precision, is_fiat, blockchain, contract_address, withdrawal_min_balance, has_balance, status, sort_order, min_confirmation, created_at, updated_at, is_stablecoin, currency_label, token_label, is_native, is_new_store_default, order_idx FROM currencies WHERE id=$1 LIMIT 1
 `
 
 func (q *Queries) GetByID(ctx context.Context, id string) (*models.Currency, error) {
@@ -152,6 +160,8 @@ func (q *Queries) GetByID(ctx context.Context, id string) (*models.Currency, err
 		&i.CurrencyLabel,
 		&i.TokenLabel,
 		&i.IsNative,
+		&i.IsNewStoreDefault,
+		&i.OrderIdx,
 	)
 	return &i, err
 }