+++ sql/postgres/migrations/20250930092613_add_currency_is_enabled_by_default.up.sql
@@ -1,0 +1,2 @@
+ALTER TABLE currencies 
+    ADD COLUMN is_new_store_default BOOLEAN NOT NULL DEFAULT FALSE;